Improve libcheck finding
This patch allows the libcheck package to provide two targets:

  - Check::LibCompat
  - Check::LibCheck

The libcompat is needed on Windows so should be defined as a target. The
dependencies are correctly defined so that the external project is built
before the libraries are used. This patch also adds the dependent
system libraries into the LIBCHECK_LIBRARIES variable if the user would
like to use that.